Ignore misplaced limits commands
Instead of rendering \limits or \nolimits when those
commands are misplaced the commands are ignored.

TeX does it the same way, but issues a warning in addition.
Issuing the warning will be taken care of in a follow-up commit.

(cherry picked from commit b058d88bf9499fa7b7fcd14c6737dde8f5709189)

Bug: T365170
Change-Id: Ic1c28be2a5669183c3dd5d15577a35a1b5a76971