Extend MathML test coverage
Render all formulae from Wikipedia and make sure no errors
(that end up as wikimedia production errors in logstash)
are thrown.
Check that no merror tags are generated for the manual page
that intends to cover the full spectrum.
Currently, some merror tags are generated, which we ignore
for this commit.

Change-Id: Ifecbd8a1b0afabaa2f818e9241add92a9508bae9