Fix rendering of \operatorname{a}'
LaTeX allows to render derivatives in combination with
operatorname in a way tha the derivative is put after
the operator and before the function application
space. This mimics that behavior as a special case.

It would be easier to construct an output XML tree
and massage the tree instead of juggling around
with state and output strings. However, that would be
a major change.

Bug: T384794
Change-Id: I80c3c314ca0a33b36f5bd993bb3786b8d5dce310