Add changes test
The native Math rendering mode is transitioning into
a more stable state. To avoid regression, we start
adding changes test to avoid unintended side effects
for modifications made. The maintenance script
provides a mechanism, to easily update the references
after manual review.

When creating the *unit* tests it was required to
move forward with refactoring the code to better
support the concept of services. Thus new service
injection was added to make the code testable by
unit code without to disable the old entry paths.

The newly introduced property, rawError is a
workaround as it calls

\MediaWiki\Message\Message::fetchMessage

which relies on MediaWikiServices to be present.

Bug: T386071
Change-Id: I52c1b23194dda41a85e28d436a62d34a19154783