Remove link to current site inside formula
There is a rare edge case where the elements in the formula may be linked to the current page.
This happens when MathWikibaseConnector->fetchPageUrl returns false, i.e., when the linked
element does not have a correct site link.
The fix adds another condition to MathNativeMML.php->addLinksToMathML, where empty links are ignored.
Also added a test to test this case.

Bug: T386760
Change-Id: Ide2fe3b9d97177902b5827eb23b4762a5544c4a0