tests: Remove redundant WANObjectCache override
Since 2020 with change Ia60cb0bc18b23, the main cache is enabled
with a hash bag by default (and reset between each test), so this
override is identical to the default. Since WANObjectCache is a
wrapper around the main cache, simplify the cases where an override
is needed, to use the setMainCache() helper instead.

Remove double checking of the fake itself. This is well-tested within
core already.

Change-Id: Ie63e972c90c9c4187c9ca349a47784ce4136b44c