HookRunner: Add new hook MathFormulaPostRenderRevision.
When math formula is rendered, MathFormulaPostRender hook is fired.
This hook is insufficient for future use with parsoid rendering,
because current implementation uses legacy parser as argument.
This argument cannot be simply extended or override, so this
patch introduces new variant, which uses RevisionRecord instead.
This patch contains hook registration and implementation
for legacy parser hooks. New hook should replace
MathFormulaPostRender hook, which is since this patch soft
deprecated.

Bug: T399226
Change-Id: Ie1def0b7cbb8cf928a44bb5e8431889049faf813